fable library w.i.p
