+++ package.json
@@ -20,6 +20,7 @@
     "@types/node": "20.11.0",
     "@types/react": "18.2.47",
     "@types/react-dom": "18.2.18",
+    "file-loader": "^6.2.0",
     "next": "14.0.4",
     "react": "^18",
     "react-cookie": "^7.1.0",
