+++ src/components/TagElement/TagElement.types.ts
@@ -1,6 +1,6 @@
 import type { ReactElement } from 'react';
 
-import type { TAG_NAMES } from '@/fables_library/library.types';
+import type { TAG_NAMES } from '@/fablesLibrary/library.types';
 
 export interface TagElementProps {
   tagName: TAG_NAMES;
