+++ src/components/TagList/TagList.types.ts
@@ -1,6 +1,6 @@
 import type { ReactElement } from 'react';
 
-import type { TAG_NAMES } from '@/fables_library/library.types';
+import type { TAG_NAMES } from '@/fablesLibrary/library.types';
 
 export interface TagListProps {
   storyTitle: string;
