+++ src/components/Tagbox/Tagbox.types.ts
@@ -1,6 +1,6 @@
 import type { Dispatch, ReactElement, SetStateAction } from 'react';
 
-import type { TAG_NAMES } from '@/fables_library/library.types';
+import type { TAG_NAMES } from '@/fablesLibrary/library.types';
 
 export interface TagboxProps {
   exitFunction: Dispatch<SetStateAction<boolean>>;
