+++ src/components/TemplateStoryPage/TemplateStoryPage.types.ts
@@ -1,6 +1,6 @@
 import type { ReactElement } from 'react';
 
-import type { Story } from '@/fables_library/library.types';
+import type { Story } from '@/fablesLibrary/library.types';
 
 export interface TemplateStoryPageProps {
   storyData: Story;
