+++ src/types/custom.d.ts
@@ -1,0 +1,6 @@
+// types/custom.d.ts
+
+declare module '*.webm' {
+  const value: string;
+  export default value;
+}
